Add an FAQ about linking / installation problems on M1 MacOSX
In short, the problem is that many people are using Emacs compiled for
x86_64 on the M1 Mac, which is an ARM64 architecture. There is nothing
that `pdf-tools` can do here, but the "fix" is straightforward:
Install Emacs from https://emacsformacosx.com/, which is compiled
correctly.

Relates to: vedang/pdf-tools#107